+++ Mcsg.Identity.Api/Setting.cs
@@ -54,12 +54,12 @@ public class Setting : SettingBase, ISetting
     /// <summary>
     /// The remaining time in minutes after which the username can be changed
     /// </summary>
-    public uint UserNameChangedInRemaining { get; set; }
+    public double UserNameChangedInRemaining { get; set; }
 
     /// <summary>
     /// The waiting time in minutes before the username can be changed
     /// </summary>
-    public uint UserNameWaitingChangedAfter { get; set; }
+    public double UserNameWaitingChangedAfter { get; set; }
 
     #endregion
